Revert "Combine the 2 test classes"
This reverts commit 70c31783dd269e038258b0f64b1f8836a2672bc8.
