+++ solr/solrj/src/test/org/apache/solr/client/solrj/impl/LBHttp2SolrClientTest.java
@@ -16,226 +16,26 @@
  */
 package org.apache.solr.client.solrj.impl;
 
-import java.io.File;
-import java.io.IOException;
-import java.lang.invoke.MethodHandles;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Properties;
 import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-
-import org.apache.lucene.util.IOUtils;
-import org.apache.solr.SolrTestCaseJ4;
-import org.apache.solr.client.solrj.SolrClient;
-import org.apache.solr.client.solrj.SolrQuery;
+import org.apache.solr.SolrTestCase;
 import org.apache.solr.client.solrj.SolrRequest;
-import org.apache.solr.client.solrj.SolrServerException;
-import org.apache.solr.client.solrj.impl.Http2SolrClient;
-import org.apache.solr.client.solrj.impl.LBHttp2SolrClient;
-import org.apache.solr.client.solrj.impl.LBSolrClient;
 import org.apache.solr.client.solrj.request.QueryRequest;
-import org.apache.solr.client.solrj.response.QueryResponse;
-import org.apache.solr.client.solrj.response.SolrResponseBase;
 import org.apache.solr.common.SolrException;
-import org.apache.solr.common.SolrInputDocument;
 import org.apache.solr.common.params.MapSolrParams;
 import org.apache.solr.common.util.NamedList;
-import org.apache.solr.common.util.TimeSource;
-import org.apache.solr.embedded.JettyConfig;
-import org.apache.solr.embedded.JettySolrRunner;
-import org.apache.solr.util.TimeOut;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-/**
- * Test for LBHttp2SolrClient
- *
- * @since solr 1.4
- */
-public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
-
-  private static final Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-
-  SolrInstance[] solr = new SolrInstance[3];
-  Http2SolrClient httpClient;
-
-  // TODO: fix this test to not require FSDirectory
-  static String savedFactory;
-
-  @BeforeClass
-  public static void beforeClass() {
-    savedFactory = System.getProperty("solr.DirectoryFactory");
-    System.setProperty("solr.directoryFactory", "org.apache.solr.core.MockFSDirectoryFactory");
-    System.setProperty("tests.shardhandler.randomSeed", Long.toString(random().nextLong()));
-  }
-
-  @AfterClass
-  public static void afterClass() {
-    if (savedFactory == null) {
-      System.clearProperty("solr.directoryFactory");
-    } else {
-      System.setProperty("solr.directoryFactory", savedFactory);
-    }
-    System.clearProperty("tests.shardhandler.randomSeed");
-  }
-
-  @Override
-  public void setUp() throws Exception {
-    super.setUp();
-    httpClient =
-        new Http2SolrClient.Builder()
-            .withConnectionTimeout(1000, TimeUnit.MILLISECONDS)
-            .withIdleTimeout(2000, TimeUnit.MILLISECONDS)
-            .build();
-
-    for (int i = 0; i < solr.length; i++) {
-      solr[i] =
-          new SolrInstance("solr/collection1" + i, createTempDir("instance-" + i).toFile(), 0);
-      solr[i].setUp();
-      solr[i].startJetty();
-      addDocs(solr[i]);
-    }
-  }
-
-  private void addDocs(SolrInstance solrInstance) throws IOException, SolrServerException {
-    List<SolrInputDocument> docs = new ArrayList<>();
-    for (int i = 0; i < 10; i++) {
-      SolrInputDocument doc = new SolrInputDocument();
-      doc.addField("id", i);
-      doc.addField("name", solrInstance.name);
-      docs.add(doc);
-    }
-    SolrResponseBase resp;
-    try (SolrClient client =
-        getHttpSolrClient(solrInstance.getBaseUrl(), solrInstance.getDefaultCollection())) {
-      resp = client.add(docs);
-      assertEquals(0, resp.getStatus());
-      resp = client.commit();
-      assertEquals(0, resp.getStatus());
-    }
-  }
-
-  @Override
-  public void tearDown() throws Exception {
-    for (SolrInstance aSolr : solr) {
-      if (aSolr != null) {
-        aSolr.tearDown();
-      }
-    }
-    httpClient.close();
-    super.tearDown();
-  }
-
-  public void testSimple() throws Exception {
-    final var baseSolrEndpoints = bootstrapBaseSolrEndpoints(solr.length);
-    try (LBHttp2SolrClient client =
-        new LBHttp2SolrClient.Builder(httpClient, baseSolrEndpoints)
-            .withDefaultCollection(solr[0].getDefaultCollection())
-            .setAliveCheckInterval(500, TimeUnit.MILLISECONDS)
-            .build()) {
-      SolrQuery solrQuery = new SolrQuery("*:*");
-      Set<String> names = new HashSet<>();
-      QueryResponse resp = null;
-      for (int i = 0; i < solr.length; i++) {
-        resp = client.query(solrQuery);
-        assertEquals(10, resp.getResults().getNumFound());
-        names.add(resp.getResults().get(0).getFieldValue("name").toString());
-      }
-      assertEquals(3, names.size());
-
-      // Kill a server and test again
-      solr[1].jetty.stop();
-      solr[1].jetty = null;
-      names.clear();
-      for (int i = 0; i < solr.length; i++) {
-        resp = client.query(solrQuery);
-        assertEquals(10, resp.getResults().getNumFound());
-        names.add(resp.getResults().get(0).getFieldValue("name").toString());
-      }
-      assertEquals(2, names.size());
-      assertFalse(names.contains("solr1"));
-
-      // Start the killed server once again
-      solr[1].startJetty();
-      // Wait for the alive check to complete
-      Thread.sleep(1200);
-      names.clear();
-      for (int i = 0; i < solr.length; i++) {
-        resp = client.query(solrQuery);
-        assertEquals(10, resp.getResults().getNumFound());
-        names.add(resp.getResults().get(0).getFieldValue("name").toString());
-      }
-      assertEquals(3, names.size());
-    }
-  }
-
-  public void testTwoServers() throws Exception {
-    final var baseSolrEndpoints = bootstrapBaseSolrEndpoints(2);
-    try (LBHttp2SolrClient client =
-        new LBHttp2SolrClient.Builder(httpClient, baseSolrEndpoints)
-            .withDefaultCollection(solr[0].getDefaultCollection())
-            .setAliveCheckInterval(500, TimeUnit.MILLISECONDS)
-            .build()) {
-      SolrQuery solrQuery = new SolrQuery("*:*");
-      QueryResponse resp = null;
-      solr[0].jetty.stop();
-      solr[0].jetty = null;
-      resp = client.query(solrQuery);
-      String name = resp.getResults().get(0).getFieldValue("name").toString();
-      assertEquals("solr/collection11", name);
-      resp = client.query(solrQuery);
-      name = resp.getResults().get(0).getFieldValue("name").toString();
-      assertEquals("solr/collection11", name);
-      solr[1].jetty.stop();
-      solr[1].jetty = null;
-      solr[0].startJetty();
-      Thread.sleep(1200);
-      try {
-        resp = client.query(solrQuery);
-      } catch (SolrServerException e) {
-        // try again after a pause in case the error is lack of time to start server
-        Thread.sleep(3000);
-        resp = client.query(solrQuery);
-      }
-      name = resp.getResults().get(0).getFieldValue("name").toString();
-      assertEquals("solr/collection10", name);
-    }
-  }
-
-  public void testReliability() throws Exception {
-    final var baseSolrEndpoints = bootstrapBaseSolrEndpoints(solr.length);
-
-    try (LBHttp2SolrClient client =
-        new LBHttp2SolrClient.Builder(httpClient, baseSolrEndpoints)
-            .withDefaultCollection(solr[0].getDefaultCollection())
-            .setAliveCheckInterval(500, TimeUnit.MILLISECONDS)
-            .build()) {
-
-      // Kill a server and test again
-      solr[1].jetty.stop();
-      solr[1].jetty = null;
-
-      // query the servers
-      for (int i = 0; i < solr.length; i++) client.query(new SolrQuery("*:*"));
-
-      // Start the killed server once again
-      solr[1].startJetty();
-      // Wait for the alive check to complete
-      waitForServer(30, client, 3, solr[1].name);
-    }
-  }
+/** Test the LBHttp2SolrClient. */
+public class LBHttp2SolrClientTest extends SolrTestCase {
 
   /**
    * Test method for {@link LBHttp2SolrClient.Builder} that validates that the query param keys
@@ -249,19 +49,19 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
     urlParamNames.add("param1");
 
     try (Http2SolrClient http2SolrClient =
-                 new Http2SolrClient.Builder(url).withTheseParamNamesInTheUrl(urlParamNames).build();
-         LBHttp2SolrClient testClient =
-                 new LBHttp2SolrClient.Builder(http2SolrClient, new LBSolrClient.Endpoint(url))
-                         .build()) {
+            new Http2SolrClient.Builder(url).withTheseParamNamesInTheUrl(urlParamNames).build();
+        LBHttp2SolrClient testClient =
+            new LBHttp2SolrClient.Builder(http2SolrClient, new LBSolrClient.Endpoint(url))
+                .build()) {
 
       assertArrayEquals(
-              "Wrong urlParamNames found in lb client.",
-              urlParamNames.toArray(),
-              testClient.getUrlParamNames().toArray());
+          "Wrong urlParamNames found in lb client.",
+          urlParamNames.toArray(),
+          testClient.getUrlParamNames().toArray());
       assertArrayEquals(
-              "Wrong urlParamNames found in base client.",
-              urlParamNames.toArray(),
-              http2SolrClient.getUrlParamNames().toArray());
+          "Wrong urlParamNames found in base client.",
+          urlParamNames.toArray(),
+          http2SolrClient.getUrlParamNames().toArray());
     }
   }
 
@@ -276,10 +76,10 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
     List<LBSolrClient.Endpoint> endpointList = List.of(ep1, ep2);
 
     Http2SolrClient.Builder b =
-            new Http2SolrClient.Builder("http://base.url").withConnectionTimeout(10, TimeUnit.SECONDS);
+        new Http2SolrClient.Builder("http://base.url").withConnectionTimeout(10, TimeUnit.SECONDS);
     ;
     try (MockHttp2SolrClient client = new MockHttp2SolrClient("http://base.url", b);
-         LBHttp2SolrClient testClient = new LBHttp2SolrClient.Builder(client, ep1, ep2).build()) {
+        LBHttp2SolrClient testClient = new LBHttp2SolrClient.Builder(client, ep1, ep2).build()) {
 
       for (int j = 0; j < 2; j++) {
         // first time Endpoint One will return error code 500.
@@ -340,9 +140,9 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
     List<LBSolrClient.Endpoint> endpointList = List.of(ep1, ep2);
 
     Http2SolrClient.Builder b =
-            new Http2SolrClient.Builder("http://base.url").withConnectionTimeout(10, TimeUnit.SECONDS);
+        new Http2SolrClient.Builder("http://base.url").withConnectionTimeout(10, TimeUnit.SECONDS);
     try (MockHttp2SolrClient client = new MockHttp2SolrClient("http://base.url", b);
-         LBHttp2SolrClient testClient = new LBHttp2SolrClient.Builder(client, ep1, ep2).build()) {
+        LBHttp2SolrClient testClient = new LBHttp2SolrClient.Builder(client, ep1, ep2).build()) {
 
       int limit = 10; // For simplicity use an even limit
       List<CompletableFuture<LBSolrClient.Rsp>> responses = new ArrayList<>();
@@ -395,126 +195,6 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
     }
   }
 
-  // wait maximum ms for serverName to come back up
-  private void waitForServer(
-          int maxSeconds, LBHttp2SolrClient client, int nServers, String serverName) throws Exception {
-    final TimeOut timeout = new TimeOut(maxSeconds, TimeUnit.SECONDS, TimeSource.NANO_TIME);
-    while (!timeout.hasTimedOut()) {
-      QueryResponse resp;
-      try {
-        resp = client.query(new SolrQuery("*:*"));
-      } catch (Exception e) {
-        log.warn("", e);
-        continue;
-      }
-      String name = resp.getResults().get(0).getFieldValue("name").toString();
-      if (name.equals(serverName)) return;
-
-      Thread.sleep(500);
-    }
-  }
-
-  private LBSolrClient.Endpoint[] bootstrapBaseSolrEndpoints(int max) {
-    LBSolrClient.Endpoint[] solrUrls = new LBSolrClient.Endpoint[max];
-    for (int i = 0; i < max; i++) {
-      solrUrls[i] = new LBSolrClient.Endpoint(solr[i].getBaseUrl());
-    }
-    return solrUrls;
-  }
-
-  private static class SolrInstance {
-    String name;
-    File homeDir;
-    File dataDir;
-    File confDir;
-    int port;
-    JettySolrRunner jetty;
-
-    public SolrInstance(String name, File homeDir, int port) {
-      this.name = name;
-      this.homeDir = homeDir;
-      this.port = port;
-
-      dataDir = new File(homeDir + "/collection1", "data");
-      confDir = new File(homeDir + "/collection1", "conf");
-    }
-
-    public String getHomeDir() {
-      return homeDir.toString();
-    }
-
-    public String getUrl() {
-      return buildUrl(port) + "/collection1";
-    }
-
-    public String getBaseUrl() {
-      return buildUrl(port);
-    }
-
-    public String getDefaultCollection() {
-      return "collection1";
-    }
-
-    public String getSchemaFile() {
-      return "solrj/solr/collection1/conf/schema-replication1.xml";
-    }
-
-    public String getConfDir() {
-      return confDir.toString();
-    }
-
-    public String getDataDir() {
-      return dataDir.toString();
-    }
-
-    public String getSolrConfigFile() {
-      return "solrj/solr/collection1/conf/solrconfig-follower1.xml";
-    }
-
-    public String getSolrXmlFile() {
-      return "solrj/solr/solr.xml";
-    }
-
-    public void setUp() throws Exception {
-      homeDir.mkdirs();
-      dataDir.mkdirs();
-      confDir.mkdirs();
-
-      Files.copy(
-          SolrTestCaseJ4.getFile(getSolrXmlFile()).toPath(), homeDir.toPath().resolve("solr.xml"));
-
-      Path f = confDir.toPath().resolve("solrconfig.xml");
-      Files.copy(SolrTestCaseJ4.getFile(getSolrConfigFile()).toPath(), f);
-      f = confDir.toPath().resolve("schema.xml");
-      Files.copy(SolrTestCaseJ4.getFile(getSchemaFile()).toPath(), f);
-      Files.createFile(homeDir.toPath().resolve("collection1/core.properties"));
-    }
-
-    public void tearDown() throws Exception {
-      if (jetty != null) jetty.stop();
-      IOUtils.rm(homeDir.toPath());
-    }
-
-    public void startJetty() throws Exception {
-
-      Properties props = new Properties();
-      props.setProperty("solrconfig", "bad_solrconfig.xml");
-      props.setProperty("solr.data.dir", getDataDir());
-
-      JettyConfig jettyConfig = JettyConfig.builder().setPort(port).build();
-
-      jetty = new JettySolrRunner(getHomeDir(), props, jettyConfig);
-      jetty.start();
-      int newPort = jetty.getLocalPort();
-      if (port != 0 && newPort != port) {
-        fail("TESTING FAILURE: could not grab requested port.");
-      }
-      this.port = newPort;
-      //      System.out.println("waiting.........");
-      //      Thread.sleep(5000);
-    }
-  }
-
   public static class MockHttp2SolrClient extends Http2SolrClient {
 
     public List<SolrRequest<?>> lastSolrRequests = new ArrayList<>();
@@ -534,14 +214,14 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
 
     @Override
     public CompletableFuture<NamedList<Object>> requestAsync(
-            final SolrRequest<?> solrRequest, String collection) {
+        final SolrRequest<?> solrRequest, String collection) {
       CompletableFuture<NamedList<Object>> cf = new CompletableFuture<>();
       lastSolrRequests.add(solrRequest);
       lastBasePaths.add(solrRequest.getBasePath());
       lastCollections.add(collection);
       if (solrRequest.getBasePath().equals(basePathToFail)) {
         cf.completeExceptionally(
-                new SolrException(SolrException.ErrorCode.SERVER_ERROR, "We should retry this."));
+            new SolrException(SolrException.ErrorCode.SERVER_ERROR, "We should retry this."));
       } else {
         cf.complete(generateResponse(solrRequest));
       }
@@ -553,5 +233,4 @@ public class LBHttp2SolrClientTest extends SolrTestCaseJ4 {
       return new NamedList<>(Collections.singletonMap("response", id));
     }
   }
-
 }
